Rodolfo

+++ Registration/src/Registration.Handlers/Handlers/FirstFruitsHanler.cs
@@ -10,13 +10,24 @@ using Registration.Mapper.DTOs.FirstFruits;
 using Registration.DomainCore.HandlerAbstraction;
 
 namespace Registration.Handlers.Handlers;
-public class FirstFruitsHanler : Handler
+public sealed class FirstFruitsHanler : Handler
 {
     private IFirstFruitsRepository _context;
-    
-    public FirstFruitsHanler(IFirstFruitsRepository context, CViewModel viewModel, IMapper mapper) : base(mapper, viewModel)
+    private OperationsHandler _operationsHandler;
+
+    public FirstFruitsHanler(IFirstFruitsRepository context, CViewModel viewModel, IMapper mapper, OperationsHandler operationsHandler) 
+        : base(mapper, viewModel)
     {
         _context = context;
+        _operationsHandler = operationsHandler;
+    }
+
+    protected override async Task<bool> MonthWorkIsBlock(string competence, int churchId)
+    {
+        var yearMonth = DateTime.Parse(competence).ToString("yyyyMM");
+        var monthWork = await _operationsHandler.GetOneByCompetence(yearMonth, churchId);
+
+        return monthWork == null ? false : true;
     }
 
     public async Task<CViewModel> GetAll(int churchId, bool active = true)
@@ -41,7 +52,7 @@ public class FirstFruitsHanler : Handler
         catch
         {
             _statusCode = (int)Scode.INTERNAL_SERVER_ERROR;
-            _viewModel.SetErrors("Internal Error - FF1101A");
+            _viewModel!.SetErrors("Internal Error - FF1101A");
         }
 
         return _viewModel;
@@ -54,7 +65,7 @@ public class FirstFruitsHanler : Handler
             if (!ValidateCompetence(competence))
             {
                 _statusCode = (int)Scode.BAD_REQUEST;
-                _viewModel.SetErrors("Request Error. Check the properties - FF1103A");
+                _viewModel!.SetErrors("Request Error. Check the properties - FF1103A");
 
                 return _viewModel;
             }
@@ -78,7 +89,7 @@ public class FirstFruitsHanler : Handler
         catch
         {
             _statusCode = (int)Scode.INTERNAL_SERVER_ERROR;
-            _viewModel.SetErrors("Internal Error - FF1103B");
+            _viewModel!.SetErrors("Internal Error - FF1103B");
         }
 
         return _viewModel;
@@ -92,7 +103,7 @@ public class FirstFruitsHanler : Handler
             if (firstFruits == null)
             {
                 _statusCode = (int)Scode.NOT_FOUND;
-                _viewModel.SetErrors("Object not found");
+                _viewModel!.SetErrors("Object not found");
 
                 return _viewModel;
             }
@@ -105,7 +116,7 @@ public class FirstFruitsHanler : Handler
         catch
         {
             _statusCode = (int)Scode.INTERNAL_SERVER_ERROR;
-            _viewModel.SetErrors("Internal Error - FF1102A");
+            _viewModel!.SetErrors("Internal Error - FF1102A");
         }
 
         return _viewModel;
@@ -117,7 +128,15 @@ public class FirstFruitsHanler : Handler
         if (!firstFruitsEditDto.IsValid)
         {
             _statusCode = (int)Scode.BAD_REQUEST;
-            _viewModel.SetErrors(firstFruitsEditDto.GetNotification());
+            _viewModel!.SetErrors(firstFruitsEditDto.GetNotification());
+
+            return _viewModel;
+        }
+
+        if(await MonthWorkIsBlock(firstFruitsEditDto.Competence, firstFruitsEditDto.ChurchId))
+        {
+            _statusCode = (int)Scode.NOT_ACCEPTABLE;
+            _viewModel!.SetErrors("This competence has already been closed!");
 
             return _viewModel;
         }
@@ -127,9 +146,9 @@ public class FirstFruitsHanler : Handler
             var firstFruits = _mapper.Map<FirstFruits>(firstFruitsEditDto);
             await _context.Post(firstFruits)!;
 
-            var newTithes = await _context.GetOne(firstFruits.Id);
+            var newFirstFruits = await _context.GetOne(firstFruits.Id);
 
-            var firstFruitsReadDto = _mapper.Map<ReadFirstFruitsDto>(newTithes);
+            var firstFruitsReadDto = _mapper.Map<ReadFirstFruitsDto>(newFirstFruits);
             _statusCode = (int)Scode.CREATED;
 
             _viewModel.SetData(firstFruitsReadDto);
@@ -137,12 +156,12 @@ public class FirstFruitsHanler : Handler
         catch (DbUpdateException)
         {
             _statusCode = (int)Scode.BAD_REQUEST;
-            _viewModel.SetErrors("Request Error. Check the properties - FF1103A");
+            _viewModel!.SetErrors("Request Error. Check the properties - FF1103A");
         }
         catch
         {
             _statusCode = (int)Scode.INTERNAL_SERVER_ERROR;
-            _viewModel.SetErrors("Internal Error - FF1103B");
+            _viewModel!.SetErrors("Internal Error - FF1103B");
         }
 
         return _viewModel;
@@ -154,7 +173,15 @@ public class FirstFruitsHanler : Handler
         if (!firstFruitsEditDto.IsValid)
         {
             _statusCode = (int)Scode.BAD_REQUEST;
-            _viewModel.SetErrors(firstFruitsEditDto.GetNotification());
+            _viewModel!.SetErrors(firstFruitsEditDto.GetNotification());
+        }
+
+        if (await MonthWorkIsBlock(firstFruitsEditDto.Competence, firstFruitsEditDto.ChurchId))
+        {
+            _statusCode = (int)Scode.NOT_ACCEPTABLE;
+            _viewModel!.SetErrors("This competence has already been closed!");
+
+            return _viewModel;
         }
 
         try
@@ -163,7 +190,7 @@ public class FirstFruitsHanler : Handler
             if (firstFruits == null)
             {
                 _statusCode = 404;
-                _viewModel.SetErrors("Object not found");
+                _viewModel!.SetErrors("Object not found");
             }
 
             var editTithes = _mapper.Map<FirstFruits>(firstFruitsEditDto);
@@ -195,7 +222,15 @@ public class FirstFruitsHanler : Handler
             if (firstFruits == null)
             {
                 _statusCode = (int)Scode.NOT_FOUND;
-                _viewModel.SetErrors("Object not found");
+                _viewModel!.SetErrors("Object not found");
+            }
+
+            if (await MonthWorkIsBlock(firstFruits.Competence, firstFruits.ChurchId))
+            {
+                _statusCode = (int)Scode.NOT_ACCEPTABLE;
+                _viewModel!.SetErrors("This competence has already been closed!");
+
+                return _viewModel;
             }
 
             await _context.Delete(firstFruits);
