+++ Registration/src/Registration.Handlers/Handlers/OutFlowHanler.cs
@@ -10,13 +10,23 @@ using Registration.DomainCore.ViewModelAbstraction;
 using Registration.DomainCore.HandlerAbstraction;
 
 namespace Registration.Handlers.Handlers;
-public class OutFlowHanler : Handler
+public sealed class OutFlowHanler : Handler
 {
     private IOutFlowRepository _context;
+    private OperationsHandler _operationsHandler;
 
-    public OutFlowHanler(IOutFlowRepository context, IMapper mapper, CViewModel viewModel) : base(mapper, viewModel)
+    public OutFlowHanler(IOutFlowRepository context, IMapper mapper, CViewModel viewModel, OperationsHandler operationsHandler) : base(mapper, viewModel)
     {
         _context = context;
+        _operationsHandler = operationsHandler;
+    }
+
+    protected override async Task<bool> MonthWorkIsBlock(string competence, int churchId)
+    {
+        var yearMonth = DateTime.Parse(competence).ToString("yyyyMM");
+        var monthWork = await _operationsHandler.GetOneByCompetence(yearMonth, churchId);
+
+        return monthWork == null ? false : true;
     }
 
     public async Task<CViewModel> GetAll(int churchId, bool active = true)
@@ -41,7 +51,7 @@ public class OutFlowHanler : Handler
         catch
         {
             _statusCode = (int)Scode.INTERNAL_SERVER_ERROR;
-            _viewModel.SetErrors("Internal Error - OT1101A");
+            _viewModel!.SetErrors("Internal Error - OT1101A");
         }
 
         return _viewModel;
@@ -55,7 +65,7 @@ public class OutFlowHanler : Handler
             if (outFlow == null)
             {
                 _statusCode = (int)Scode.NOT_FOUND;
-                _viewModel.SetErrors("Object not found");
+                _viewModel!.SetErrors("Object not found");
 
                 return _viewModel;
             }
@@ -68,7 +78,7 @@ public class OutFlowHanler : Handler
         catch
         {
             _statusCode = (int)Scode.INTERNAL_SERVER_ERROR;
-            _viewModel.SetErrors("Internal Error - OT1102A");
+            _viewModel!.SetErrors("Internal Error - OT1102A");
         }
 
         return _viewModel;
@@ -80,7 +90,15 @@ public class OutFlowHanler : Handler
         if (!outFlowEditDto.IsValid)
         {
             _statusCode = (int)Scode.BAD_REQUEST;
-            _viewModel.SetErrors(outFlowEditDto.GetNotification());
+            _viewModel!.SetErrors(outFlowEditDto.GetNotification());
+
+            return _viewModel;
+        }
+
+        if (await MonthWorkIsBlock(outFlowEditDto.Competence, outFlowEditDto.ChurchId))
+        {
+            _statusCode = (int)Scode.NOT_ACCEPTABLE;
+            _viewModel!.SetErrors("This competence has already been closed!");
 
             return _viewModel;
         }
@@ -103,12 +121,12 @@ public class OutFlowHanler : Handler
         catch (DbUpdateException)
         {
             _statusCode = (int)Scode.BAD_REQUEST;
-            _viewModel.SetErrors("Request Error. Check the properties - OT1103A");
+            _viewModel!.SetErrors("Request Error. Check the properties - OT1103A");
         }
         catch(Exception ex)
         {
             _statusCode = (int)Scode.INTERNAL_SERVER_ERROR;
-            _viewModel.SetErrors("Internal Error. - OT1103B");
+            _viewModel!.SetErrors("Internal Error. - OT1103B");
         }
 
         return _viewModel;
@@ -120,7 +138,16 @@ public class OutFlowHanler : Handler
         if (!outFlowEditDto.IsValid)
         {
             _statusCode = (int)Scode.BAD_REQUEST;
-            _viewModel.SetErrors(outFlowEditDto.GetNotification());
+            _viewModel!.SetErrors(outFlowEditDto.GetNotification());
+        }
+
+
+        if (await MonthWorkIsBlock(outFlowEditDto.Competence, outFlowEditDto.ChurchId))
+        {
+            _statusCode = (int)Scode.NOT_ACCEPTABLE;
+            _viewModel!.SetErrors("This competence has already been closed!");
+
+            return _viewModel;
         }
 
         try
@@ -129,7 +156,7 @@ public class OutFlowHanler : Handler
             if (outFlow == null)
             {
                 _statusCode = 404;
-                _viewModel.SetErrors("Object not found");
+                _viewModel!.SetErrors("Object not found");
             }
 
             var editOutFlow = _mapper.Map<OutFlow>(outFlowEditDto);
@@ -163,7 +190,15 @@ public class OutFlowHanler : Handler
             if (otFlow == null)
             {
                 _statusCode = (int)Scode.NOT_FOUND;
-                _viewModel.SetErrors("Object not found");
+                _viewModel!.SetErrors("Object not found");
+            }
+
+            if (await MonthWorkIsBlock(otFlow.Competence, otFlow.ChurchId))
+            {
+                _statusCode = (int)Scode.NOT_ACCEPTABLE;
+                _viewModel!.SetErrors("This competence has already been closed!");
+
+                return _viewModel;
             }
 
             await _context.Delete(otFlow);
